requirements.txt
torch>=2.0
torchvision>=0.15
albumentations>=1.3.0
timm>=0.9.0
opencv-python-headless>=4.5
matplotlib>=3.5
seaborn>=0.12
scikit-learn>=1.2
scipy>=1.9
numpy>=1.23
pandas>=1.5
Pillow>=9.0
tqdm>=4.64
PyYAML>=6.0
tensorboard>=2.11
fastapi>=0.110
uvicorn>=0.27
python-multipart>=0.0.9
